fake font measurements
