+++ packages/cursorless-vscode-core/src/index.ts
@@ -2,5 +2,7 @@ export * from "./StatusBarItem";
 export * from "./commands";
 export * from "./ide/vscode/VscodeIDE";
 export * from "./ide/vscode/hats/VscodeHats";
+export * from "./ide/vscode/hats/FontMeasurementsImpl";
+export * from "./ide/vscode/hats/FakeFontMeasurements";
 export * from "./ide/vscode/toVscodeEditor";
 export * from "./keyboard/KeyboardCommands";
